Move BranchRestorer to app.rs
