+++ src/git.rs
@@ -49,33 +49,6 @@ pub enum GitError {
     UnexpectedOutput(String),
 }
 
-/**
- * Restores the current git branch when dropped
- * Assumes we are on a real branch
- */
-pub struct BranchRestorer<'a> {
-    repository: &'a Repository,
-    branch: String,
-}
-
-impl BranchRestorer<'_> {
-    pub fn new(repo: &Repository) -> BranchRestorer<'_> {
-        let current_branch = repo.get_current_branch().expect("Can't get current branch");
-        BranchRestorer {
-            repository: repo,
-            branch: current_branch,
-        }
-    }
-}
-
-impl Drop for BranchRestorer<'_> {
-    fn drop(&mut self) {
-        if let Err(_x) = self.repository.checkout(&self.branch) {
-            println!("Failed to restore original branch {}", self.branch);
-        }
-    }
-}
-
 pub struct Repository {
     pub path: PathBuf,
 }
